+++ src/Unchase.Swashbuckle.AspNetCore.Extensions/Filters/XmlCommentsWithRemarksParameterFilter.cs
@@ -10,7 +10,7 @@ namespace Unchase.Swashbuckle.AspNetCore.Extensions.Filters
     /// <summary>
     /// Inject human-friendly remarks to descriptions for Parameters based on XML Comment files.
     /// </summary>
-    public class XmlCommentsWithRemarksParameterFilter : IParameterFilter
+    internal class XmlCommentsWithRemarksParameterFilter : IParameterFilter
     {
         #region Fields
 
@@ -53,8 +53,7 @@ namespace Unchase.Swashbuckle.AspNetCore.Extensions.Filters
 
         private void ApplyPropertyTags(OpenApiParameter parameter, PropertyInfo propertyInfo)
         {
-            if (propertyInfo.DeclaringType != null && _excludedTypes.ToList().Select(t => t.FullName)
-                .Contains(propertyInfo.DeclaringType?.FullName))
+            if (propertyInfo.DeclaringType != null && _excludedTypes.Any() && _excludedTypes.ToList().Contains(propertyInfo.DeclaringType))
             {
                 return;
             }