+++ test/WebApi3.1-Swashbuckle/Models/IInheritDocClass.cs
@@ -1,0 +1,19 @@
+﻿namespace WebApi3._1_Swashbuckle.Models
+{
+    /// <summary>
+    /// InheritDocClass - inheritdoc
+    /// </summary>
+    /// <remarks>
+    /// InheritDocClass remarks - inheritdoc
+    /// </remarks>
+    public interface IInheritDocClass : IInheritDocCommon
+    {
+        /// <summary>
+        /// Name - inheritdoc
+        /// </summary>
+        /// <remarks>
+        /// Name remarks - inheritdoc
+        /// </remarks>
+        public string Name { get; set; }
+    }
+}