+++ test/WebApi3.1-Swashbuckle/Models/InheritDocClass.cs
@@ -1,0 +1,15 @@
+﻿namespace WebApi3._1_Swashbuckle.Models
+{
+    /// <inheritdoc cref="IInheritDocClass"/>
+    public class InheritDocClass : IInheritDocClass
+    {
+        /// <inheritdoc/>
+        public string Name { get; set; }
+
+        /// <inheritdoc/>
+        public string Common { get; set; }
+
+        /// <inheritdoc/>
+        public InheritEnum InheritEnum { get; set; }
+    }
+}