+++ test/WebApi3.1-Swashbuckle/Models/InheritEnum.cs
@@ -1,0 +1,19 @@
+﻿namespace WebApi3._1_Swashbuckle.Models
+{
+    /// <summary>
+    /// Inherit enum - enum
+    /// </summary>
+    /// <remarks>
+    /// Inherit enum remarks - enum
+    /// </remarks>
+    public enum InheritEnum
+    {
+        /// <summary>
+        /// Inherit enum Value
+        /// </summary>
+        /// <remarks>
+        /// Inherit enum Value remarks
+        /// </remarks>
+        Value = 0
+    }
+}