+++ test/WebApi3.1-Swashbuckle/Startup.cs
@@ -8,6 +8,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Net;
+using System.Runtime.InteropServices;
 using System.Text.Json.Serialization;
 using Unchase.Swashbuckle.AspNetCore.Extensions.Extensions;
 using Unchase.Swashbuckle.AspNetCore.Extensions.Filters;
@@ -34,17 +35,21 @@ namespace WebApi3._1_Swashbuckle
                 // use it if you want to hide Paths and Definitions from OpenApi documentation correctly
                 options.UseAllOfToExtendReferenceSchemas();
 
-                #region AddEnumsWithValuesFixFilters
+                // if you want to add xml comments from inheritdocs (from summary and remarks) into the swagger documentation, add:
+                // you can exclude remarks for concrete types
+                options.IncludeXmlCommentsFromInheritDocs(includeRemarks: true, excludedTypes: typeof(string));
 
                 // if you want to add xml comments from summary and remarks into the swagger documentation, first of all add:
                 // you can exclude remarks for concrete types
                 var xmlFilePath = Path.Combine(AppContext.BaseDirectory, "WebApi3.1-Swashbuckle.xml");
-                options.IncludeXmlCommentsWithRemarks(xmlFilePath, false,
+                options.IncludeXmlCommentsWithRemarks(filePath: xmlFilePath, includeControllerXmlComments: false,
                     typeof(ComplicatedClass),
                     typeof(InnerEnum));
 
                 // or add without remarks
                 //options.IncludeXmlComments(xmlFilePath);
+
+                #region AddEnumsWithValuesFixFilters
 
                 // Add filters to fix enums
                 // use by default: